🎨  Move out core logic to its own file for testable
